requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.9
botocore==1.38.9
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dlib==19.24.6
docopt==0.6.2
docutils==0.19
dotenv==0.9.9
efficientnet_pytorch==0.7.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
fvcore==0.1.5.post20221221
gast==0.6.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gotrue==2.12.0
grad-cam==1.5.5
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iniconfig==2.1.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.12.3
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lmdb==1.6.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
mss==10.0.0
multidict==6.4.3
namex==0.0.9
narwhals==1.36.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==3.1.1
postgrest==1.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
realtime==2.4.3
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rsa==4.7.2
s3transfer==0.12.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.0
shapely==2.1.0
simplejson==3.20.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
storage3==0.11.3
streamlit==1.44.1
streamlit-cookies-manager==0.2.0
StrEnum==0.4.15
stringzilla==3.12.3
supabase==2.15.1
supafunc==0.9.4
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow_keras==0.1
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tinycss2==1.4.0
toml==0.10.2
torchinfo==1.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
ttach==0.0.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yacs==0.1.8
yarg==0.1.9
yarl==1.20.0
